Ensure murfey.bootstrap is available
This one is a bit special. The directory does not contain an
'__init__.py' as it does not need to be importable, but it should be
installed as a package and hence available via 'python -m
murfey.bootstrap'. Clients will depend on this code being present in the
wheel file, and run it directly from there.