Properly close down websocket connection
Also need to stop the feeder thread. We can do this by adding a None
value to the queue and ignoring None values in the feeder thread. This
then allows a check against _alive and therefore allows the thread to
terminate. Wait for the thread to terminate in the close() call.