Add method of performing various data collection related ISPyB inserts (#69)
This is done in by the server so that IDs can be sent back to the client via websocket and can be used when triggering processing. A new thread is started on the server which listens for messages from a dedicated rabbitmq queue and performs the requested inserts. This way the ID of the inserted record is directly available to the server, but by using rabbitmq the individual messages are still resilient to failures.

Pydantic validators are used to register callback functions for changes made to an instance state in the client, triggered by messages arriving via websocket.

Co-authored-by: Anna Horstmann <anna.horstmann@diamond.ac.uk>