Add demo server API and fix tomography alignment triggering (#86)
Adding the --demo flag to murfey.server will use a demo version of the API which avoids use of ISPyB and RabbitMQ. This is done through refactoring of the FastAPI app to use FastAPI APIRouters.

Also applies some fixes to the flushing of ProcessingJob and AutoProcProgram insert and preprocessing requests. This should fix some race conditions related to the appearance of processingJobIds and autoprocProgramIds for tilt series, required to trigger processing.