Automatically detect superres data collections and use calibrated pixel sizes (#134)
If the detector is capable of collecting in superresolution then `superres` should be set to `True` in the machine configuration. If the data are then unbinned (as determined from mdoc files) the data has been collected in superresolution. 

Allow calibrated pixel sizes based on magnification to be used rather than the pixel size direct from the mdoc.