Murfey database and modular SPA processing support (#183)
Large changes to start using a PostgreSQL database to store ephemeral data about the current processing. This is used to provide support to the complicated SPA pipeline which has been broken into separate services in work elsewhere (DiamondLightSource/python-relion#190). Further work is required to make tomography and SPA processing behave the same and to make significant improvements to maintainability.

---------

Co-authored-by: stephen-riggs <stephen.riggs@diamond.ac.uk>