Make transfer monitoring through Prometheus more accurate (#196)
Increment the transferred file count and volume on transfer of every individual file in Prometheus. This wasn't done before because of concerns over large numbers of rapid connections to the Murfey database. We are keeping the transfer registration in the Murfey database batched but hopefully Prometheus can handle the throughput better. Closes #193