Allow the URL of the server hosting the rsync daemon to be separately configured (#419)
Currently the server hosting the rsync daemon used for transfer is always assumed to be the same as that hosting the Murfey server. It is possible that these would be two different hosts which would then require two URLs. Here an rsync_url configuration option is added. If left at its default value of "" the Murfey server URL will be used, otherwise the configured value is used.