Change how we use the rsync module specified in the machine configuration (#476)
Currently we assume that the rsync module name forms part of the path but this isn't necessarily the case. We should specify the whole rsync module path as the rsync_basepath in the configuration and insert the module name directly where we need it (at the point of construction of the rsync command).