Remove the client-side stashing for tomography (#473)
The client still stores a stash of data collection and processing job calls for tomography before the first mdoc is seen.
However, the server already has a preprocess stash for tomography, so the client one is not needed any more.

This removes the client stash, and just sends all requests on to the server.