Add information about the analysers into the rsyncer api calls (#571)
This adds the length and status of any corresponding analyser queue to the information returned about an rsyncer.

I've ended up breaking the one-to-one relation between endpoints in server/api/instrument.py and instrument_server/api.py, but this seemed the easiest way to combine the rsyncer and analyser information.