Remove the manual tilt offset from the user parameters, and the table which contains it (#572)
We had a `TomographyProcessingParameters` table in the database which just contained the tilt offset. This removes that table and renames the `TomographyPreprocessingParameters` to just be `TomographyProcessingParameters`.

I've also removed `manual_tilt_offset` from the client side entirely, as it is calculated server-side. I expect that will mean it disappears from the TUI display.

The renaming in particular is a high-risk change, which needs a test build and run before we can say this works