Remove need for 'ISPYB_CREDENTIALS' environment variable (#588)
'ispyb.sqlalchemy.url()' takes credentials as an argument, through which the config file can be passed directly to the function instead of relying on the 'ISPYB_CREDENTIALS' environment variable. This eliminates the last traces of our repo's need to explicitly define 'ISPYB_CREDENTIALS'.

Additionally, elements of the CI workflow and test code related to the Murfey and ISPyB databases have also been modified. 'ISPYB_CREDENTIALS' has also been dropped successfully from the test workflow, and fixtures have been put in place to allow for the creation of databases that will be reset to a desired initial state at the end of every test. This means that the database does not have to be shut down and rebooted with every database-related test, which should help with our testing times as we continue to improve Murfey's test coverage.

NOTE: Within our code base, it looks like ISPyB spawns SQLAlchemy Session objects, whereas the Murfey database spawns SQLModel ones. As such, SQLModel's functions might not work when querying the ISPyB database. The database session fixtures for the ISPyB and Murfey test databases have been written to reflect this.