Remove the assumption of a raw directory for processed outputs (#600)
Resolves one of the doppio-live issues, where we assume a raw directory for SPA.

This instead makes it a configurable option, where the raw is added to the path if process_multiple_datasets is True, so that in cases such as eBIC where we send multiple datasets to one location they are still divided up.

Also makes the path construction consistent between SPA and tomo, which will need checking.