Read and insert tomography atlas and search map locations (#616)
Registers the locations of SearchMaps and BatchPositions for tomography data collection.
This requires reading the SearchMap.dm and .xml files, and the BatchPositionsList.xml, plus finding the atlas in the Session.dm

The logic for converting from stage coordinates to pixel positions is non-trivial and does not work perfectly.
For the Krios microscopes with FALCON and K3_FLIPX cameras it appears to give reasonably accurate values (with some ambiguity over whether "zero" is top-left or bottom-left which may need fixing), but for the Talos with K3_FLIPY it does not currently work.

The Murfey database has a new table for SearchMaps and the TiltSeries table now has positions. For ispyb the SearchMaps are inserted as GridSquares. TiltSeries location inserts will be done as part of the em-tomo-align recipe.