Scale search map dimensions and location to be those on jpg (#635)
The size and location of the search maps we insert into ispyb are currently those on the atlas mrc. This scales them to be those on the jpg, and also scales the location of tomograms on search maps.

Also inserts an angle of 0 and handles inserting atlas pixel size when a data collection group is first made.

Something similar will probably be needed for the tomogram sizes on search maps, but that may need to be done in pato as there is an extra scaling based on pixel sizes as well.