Remove serialem code and revise context determination (#632)
The SerialEM code has never been used, and if we want to support SerialEM in future it would be better to redo it. This removes all code relating to it.

Also improves and tests the determination of the context in the analyser.

There are some potential difficulties with distinguishing data and metadata. The biggest problem is mdoc files which appear with the same name in both folders.

---------

Co-authored-by: Eu Pin Tien <eu-pin.tien@diamond.ac.uk>