Update command line method for reposting failed calls (#652)
Updates the cli tool used for reposting failed calls, so that it works by calling functions directly.

Also updates the router spec and changes the tool for generating that to add session_id where needed. Problems seemed to occur for the two instances where the argument type is Union, which may need fixing in future.

Needs to be merged after #651