Changes to enable transfer and processing of Tomo5 multigrid data (#662)
Multigrid tomo metadata is structured with an intermediate layer:

- session directory
  - sample directories (named Sample* as with atlas directories)
      - tomo metadata
The corresponding fractions directories are then named {session directory}_{Sample*}

Here we attempt to deal with this at the earliest possible stage, starting metadata rsyncers for the Sample* directories and hoping that everything else follows. The determination of the correct fractions directory obviously must also change for this case.