Allow database override for reinjecting failed API calls (#660)
The database is set to default to the one in the security configuration in parts of the feedback process.
To make it easier to use the API call reinjection these should take the database as input instead, and only the overall feedback_callback function should have the default.

A similar problem occurs with the machine configuration which will need to be set when using the API reinjection for now.

One naming issue is we import the database spec as db into feedback.py then have _db as a session object. This might be confusing.