Reset tilt series lengths on repeated tilts, to avoid repeat processing (#674)
When a tilt series is cancelled on the microscope, an mdoc gets written and we mark the tilt series as complete.
The current behaviour is then that we process the tomogram and do so again for every newly found tilt.

This PR changes the code to remove knowledge of tilt series length when a new tilt is found (i.e. when length exceeds the number in the mdoc). This then means we will not launch processing again until the mdoc is re-written with the correct length.

Should reduce the number of processing times from ntilts+1 to 2.
Will still need some work on the processing side as the processed star files contain duplicated tomograms