Raise an exception if no 2D murfey ids present (#683)
We keep hitting a case where no ids get sent to 2D classification. It's not clear why this happens, but if we don't have ids to send it would be better to just raise an exception in murfey.

Also I've changed the fixed number of 50 ids to the value in the default parameters (which is also 50 currently).