Feedback from tomogram picking (#682)
Ability to receive messages back from tomogram picking and then send on a request to run tomogram 2d extraction/classification.

    Registers new pj recipe for the first tomogram
    Changes SPAFeedbackParameters to a more general ClassificationFeedbackParameters for reuse
    Add table for tomogram picks

Currently then set to run 2D classification in batches of 10000 particles with 5 classes. Needs deciding if that is the right value.
Also probably isn't sending exactly the right set of parameters onwards to the next recipe as the recipe hasn't yet been written, but the right items should be stored in the database here.