Update supporting files (#712)
Add Ruff cache to the .gitignore file list.
Introduce a .dockerignore file list to prevent copying of non-essential files when building Docker containers.
Add an empty installers folder in repo for developers to store offline installers in when building Docker containers.
Add logic to murfey-server Dockerfile to check for an IMOD installation file before attempting to download it.