Replace client websocket (#716)
Replaced Client-Side Websocket Connection-Based Logging with HTTPX One
The websocket connection between the client and the backend server has been unstable, prone to error, and unable to fulfil its role of sending logs to the backend for subsequent forwarding to the logging dashboard. In this PR, the log forwarding logic has been replaced by a HTTPX-based log handler that sends batches of logs to a backend FastAPI endpoint. It also removes the now-defunct websocket-related components from the client end.

Simplifcation of pyproject.toml Dependency Keys
Consequently, the optional dependency keys needed for our pyproject.toml file can also been streamlined. As of this PR, the client and instrument-server optional dependency keys are no longer needed, as the packages needed to run the instrument server are a subset of those needed to run the backend server, and have thus been integrated into the essential dependencies section. The documentation and GitHub workflows have thus been updated to reflect this.